cargo tree transformer class refactor
